Add a new option, {maps_key_type, atom | binary}
The {maps_key_type, atom} is (still) the default.

Teh {maps_key_type, binary} only applies when the maps option is also
specified, and makes the map keys (ie the field names) be binaries
instead of atoms:

Instead of:

     #{a => 1, b => 2},

...map keys become:

     #{<<"a">> => 1, <<"b">> => 2}

Keys in map<_,_> type fields are not affected.

The purpose is easier interop with other libraries where keys are
binaries.